move 'search-bar' folder out of components, renane to more descriptive 'mock-data'
