+++ src/mock-data/MockMenteeStudent.json
@@ -1,0 +1,6 @@
+[{"id": 1, "user":{"first_name": "Rob", "last_name": "Stewart", "date_of_birth":"$date", "time_zone": "", "initials": "RS", "personal_email": "hello@email.com", "phone": "1324568695", "city": "Austin", "notes": ""}, "created_date": "$date/time", "modified_date": "$date/time", "external_id": "value", "school_level":2, "group_name": "what" },
+    {"id": 2, "user":{"first_name": "Alan", "last_name": "Davis", "date_of_birth":"$date", "time_zone": "", "initials": "AD", "personal_email": "hey@email.com", "phone": "1324568695", "city": "Austin", "notes": ""}, "created_date": "$date/time", "modified_date": "$date/time", "external_id": "value", "school_level":2, "group_name": "what" },
+    {"id": 3, "user":{"first_name": "Stacy", "last_name": "White", "date_of_birth":"$date", "time_zone": "", "initials": "RW", "personal_email": "hello@email.com", "phone": "1324568695", "city": "Austin", "notes": ""}, "created_date": "$date/time", "modified_date": "$date/time", "external_id": "value", "school_level":2, "group_name": "what" },
+    {"id": 4, "user":{"first_name": "Troy", "last_name": "Bick", "date_of_birth":"$date", "time_zone": "", "initials": "SB", "personal_email": "hello@email.com", "phone": "1324568695", "city": "Austin", "notes": ""}, "created_date": "$date/time", "modified_date": "$date/time", "external_id": "value", "school_level":2, "group_name": "what" },
+    {"id": 5, "user":{"first_name": "Rebba", "last_name": "Stack", "date_of_birth":"$date", "time_zone": "", "initials": "AS", "personal_email": "hello@email.com", "phone": "1324568695", "city": "Austin", "notes": ""}, "created_date": "$date/time", "modified_date": "$date/time", "external_id": "value", "school_level":2, "group_name": "what" },
+    {"id": 6, "user":{"first_name": "Abraar", "last_name": "Python", "date_of_birth":"$date", "time_zone": "", "initials": "HP", "personal_email": "hello@email.com", "phone": "1324568695", "city": "Austin", "notes": ""}, "created_date": "$date/time", "modified_date": "$date/time", "external_id": "value", "school_level":2, "group_name": "what" }]
