+++ src/mock-data/MockMentor.json
@@ -1,0 +1,8 @@
+[
+{"id": 7, "user":{"first_name": "Tammy", "last_name": "Nicholes", "date_of_birth":"$date", "time_zone": "", "initials": "TN", "personal_email": "hello@email.com", "phone": "1324568695", "city": "Cleveland", "notes": ""}, "created_date": "$date/time", "modified_date": "$date/time", "isInterested": true, "isIncomplete": false, "follow_up": true, "occupation": "doctor", "vbb_chapter": "one", "affiliation": "", "isinCorperateEmployeeProgram": true, "referral_source": "", "isStaff": false, "is_adult": true, "terms_agreement": true, "mentor_application_video_link": "www.heyhey.com", "application_submitted": true, "onetime_donated": false, "recurring_donation": false, "legal_notes_bymentor": false, "legal_notes_byreviewer": false, "vetted": true, "trainingScheduled": true, "attended_training": true, "trainingNotes": "very good", "completed_trainingModules": true, "nextLeadMentor_meeting_date":"right now", "nextLeadMentor_meetingInfo": "what you need to know", "leadmentor_notes": "more notes", "metHeadmaster": true, "headmaster_notes": "Not good", "metMentorAdvisor": true, "mentorAdvisor_notes": false, "additional_involvement": "nope" },
+{"id": 8, "user":{"first_name": "Heather", "last_name": "Beckman", "date_of_birth":"$date", "time_zone": "", "initials": "HB", "personal_email": "hello@email.com", "phone": "1324568695", "city": "Cleveland", "notes": ""}, "created_date": "$date/time", "modified_date": "$date/time", "isInterested": true, "isIncomplete": false, "follow_up": true, "occupation": "doctor", "vbb_chapter": "one", "affiliation": "", "isinCorperateEmployeeProgram": true, "referral_source": "", "isStaff": false, "is_adult": true, "terms_agreement": true, "mentor_application_video_link": "www.heyhey.com", "application_submitted": true, "onetime_donated": false, "recurring_donation": false, "legal_notes_bymentor": false, "legal_notes_byreviewer": false, "vetted": true, "trainingScheduled": true, "attended_training": true, "trainingNotes": "very good", "completed_trainingModules": true, "nextLeadMentor_meeting_date":"right now", "nextLeadMentor_meetingInfo": "what you need to know", "leadmentor_notes": "more notes", "metHeadmaster": true, "headmaster_notes": "Not good", "metMentorAdvisor": true, "mentorAdvisor_notes": false, "additional_involvement": "nope" },
+{"id": 9, "user":{"first_name": "Anna", "last_name": "Hangstefer", "date_of_birth":"$date", "time_zone": "", "initials": "AH", "personal_email": "hello@email.com", "phone": "1324568695", "city": "Cleveland", "notes": ""}, "created_date": "$date/time", "modified_date": "$date/time", "isInterested": true, "isIncomplete": false, "follow_up": true, "occupation": "doctor", "vbb_chapter": "one", "affiliation": "", "isinCorperateEmployeeProgram": true, "referral_source": "", "isStaff": false, "is_adult": true, "terms_agreement": true, "mentor_application_video_link": "www.heyhey.com", "application_submitted": true, "onetime_donated": false, "recurring_donation": false, "legal_notes_bymentor": false, "legal_notes_byreviewer": false, "vetted": true, "trainingScheduled": true, "attended_training": true, "trainingNotes": "very good", "completed_trainingModules": true, "nextLeadMentor_meeting_date":"right now", "nextLeadMentor_meetingInfo": "what you need to know", "leadmentor_notes": "more notes", "metHeadmaster": true, "headmaster_notes": "Not good", "metMentorAdvisor": true, "mentorAdvisor_notes": false, "additional_involvement": "nope" },
+{"id": 10, "user":{"first_name": "Germaine", "last_name": "Stephens", "date_of_birth":"$date", "time_zone": "", "initials": "GS", "personal_email": "hello@email.com", "phone": "1324568695", "city": "Cleveland", "notes": ""}, "created_date": "$date/time", "modified_date": "$date/time", "isInterested": true, "isIncomplete": false, "follow_up": true, "occupation": "doctor", "vbb_chapter": "one", "affiliation": "", "isinCorperateEmployeeProgram": true, "referral_source": "", "isStaff": false, "is_adult": true, "terms_agreement": true, "mentor_application_video_link": "www.heyhey.com", "application_submitted": true, "onetime_donated": false, "recurring_donation": false, "legal_notes_bymentor": false, "legal_notes_byreviewer": false, "vetted": true, "trainingScheduled": true, "attended_training": true, "trainingNotes": "very good", "completed_trainingModules": true, "nextLeadMentor_meeting_date":"right now", "nextLeadMentor_meetingInfo": "what you need to know", "leadmentor_notes": "more notes", "metHeadmaster": true, "headmaster_notes": "Not good", "metMentorAdvisor": true, "mentorAdvisor_notes": false, "additional_involvement": "nope" },
+{"id": 11, "user":{"first_name": "Jose", "last_name": "Smith", "date_of_birth":"$date", "time_zone": "", "initials": "JS", "personal_email": "hello@email.com", "phone": "1324568695", "city": "Cleveland", "notes": ""}, "created_date": "$date/time", "modified_date": "$date/time", "isInterested": true, "isIncomplete": false, "follow_up": true, "occupation": "doctor", "vbb_chapter": "one", "affiliation": "", "isinCorperateEmployeeProgram": true, "referral_source": "", "isStaff": false, "is_adult": true, "terms_agreement": true, "mentor_application_video_link": "www.heyhey.com", "application_submitted": true, "onetime_donated": false, "recurring_donation": false, "legal_notes_bymentor": false, "legal_notes_byreviewer": false, "vetted": true, "trainingScheduled": true, "attended_training": true, "trainingNotes": "very good", "completed_trainingModules": true, "nextLeadMentor_meeting_date":"right now", "nextLeadMentor_meetingInfo": "what you need to know", "leadmentor_notes": "more notes", "metHeadmaster": true, "headmaster_notes": "Not good", "metMentorAdvisor": true, "mentorAdvisor_notes": false, "additional_involvement": "nope" },
+{"id": 12, "user":{"first_name": "Drew", "last_name": "Statton", "date_of_birth":"$date", "time_zone": "", "initials": "DS", "personal_email": "hello@email.com", "phone": "1324568695", "city": "Cleveland", "notes": ""}, "created_date": "$date/time", "modified_date": "$date/time", "isInterested": true, "isIncomplete": false, "follow_up": true, "occupation": "doctor", "vbb_chapter": "one", "affiliation": "", "isinCorperateEmployeeProgram": true, "referral_source": "", "isStaff": false, "is_adult": true, "terms_agreement": true, "mentor_application_video_link": "www.heyhey.com", "application_submitted": true, "onetime_donated": false, "recurring_donation": false, "legal_notes_bymentor": false, "legal_notes_byreviewer": false, "vetted": true, "trainingScheduled": true, "attended_training": true, "trainingNotes": "very good", "completed_trainingModules": true, "nextLeadMentor_meeting_date":"right now", "nextLeadMentor_meetingInfo": "what you need to know", "leadmentor_notes": "more notes", "metHeadmaster": true, "headmaster_notes": "Not good", "metMentorAdvisor": true, "mentorAdvisor_notes": false, "additional_involvement": "nope" }
+]
