+++ src/mock-data/alterMockData.js
@@ -13,3 +13,5 @@ fs.writeFileSync(
   path.resolve(__dirname, './MOCK_DATA.json'),
   JSON.stringify(newData, null, 4)
 );
+
+fs.close();
